+++ heat/core/linalg/__init__.py
@@ -1,3 +1,2 @@
 from .basics import *
 from .qr import *
-from .tiling import *
